Annotations and moving of client functions
