+++ apps/web/src/components/shared/LeaveTeamDialog.tsx
@@ -12,9 +12,13 @@ import {
 } from "@/components/ui/alert-dialog";
 import { DropdownMenuItem } from "../ui/dropdown-menu";
 import { UserX } from "lucide-react";
-import { leaveTeamMutationClient as _unusedForNowleaveTeamMutationClientleaveTeamMutationClient } from "@/lib/queries";
+import { leaveTeamMutationClient as _unusedForNowleaveTeamMutationClientleaveTeamMutationClient } from "@/lib/functions/queries";
 import { useMutation as _unusedForNowuseMutation } from "@tanstack/react-query";
-
+/**
+ * Dialog component that allows a user to leave a team. It prompts the user for confirmation before proceeding with the action. If the team is private, it warns the user that they will need a new invitation to rejoin. If the team is public, it informs the user that they can rejoin later but will lose access to team data.
+ * @param param
+ * @returns 
+ */
 export function LeaveTeamDialog({
 	teamId: _unusedForNowTeamId,
 	teamName,
