+++ apps/web/src/components/shared/Navbar/UserButton.tsx
@@ -13,9 +13,9 @@ import {
 	DropdownMenuTrigger,
 } from "@/components/ui/dropdown-menu";
 import { useQuery } from "@tanstack/react-query";
-import { getUserQueryClient, getUserTeamsQueryClient } from "@/lib/queries";
+import { getUserQueryClient, getUserTeamsQueryClient } from "@/lib/functions/queries";
 import { Skeleton } from "@/components/ui/skeleton";
-import { getInitials } from "@/lib/utils";
+import { getInitials } from "@/lib/functions/utils";
 import { Plus, UserPlus } from "lucide-react";
 import { Link } from "@tanstack/react-router";
 import { authClient } from "@/lib/auth-client";
