+++ apps/web/src/components/ui/alert-dialog.tsx
@@ -1,7 +1,7 @@
 import * as React from "react";
 import * as AlertDialogPrimitive from "@radix-ui/react-alert-dialog";
 
-import { cn } from "@/lib/utils";
+import { cn } from "@/lib/functions/utils";
 import { buttonVariants } from "@/components/ui/button";
 
 function AlertDialog({
